- encapsulated related routes and handlers into single routers and controllers files
