--- backend/controllers/authController.ts
@@ -1,5 +1,0 @@
-import { Request, Response } from "express";
-
-export const authController = (req: Request, res: Response) => {
-	res.send("Auth Route");
-};
