--- backend/controllers/logoutController.ts
@@ -1,5 +1,0 @@
-import { Request, Response } from "express";
-
-export const logoutController = (req: Request, res: Response) => {
-	res.send("Register Route");
-};
