--- backend/controllers/refreshController.ts
@@ -1,5 +1,0 @@
-import { Request, Response } from "express";
-
-export const refreshController = (req: Request, res: Response) => {
-	res.send("Refresh Route");
-};
