--- backend/controllers/rootController.ts
@@ -1,5 +1,0 @@
-import { Request, Response } from "express";
-
-export const rootController = (req: Request, res: Response) => {
-	res.send("Root Route");
-};
