--- backend/routes/auth.routes.ts
@@ -1,8 +1,0 @@
-import express from "express";
-import { authController } from "../controllers/authController";
-
-const router = express.Router();
-
-router.post("/", authController);
-
-export default router;
