--- backend/routes/logout.routes.ts
@@ -1,8 +1,0 @@
-import express from "express";
-import { logoutController } from "../controllers/logoutController";
-
-const router = express.Router();
-
-router.get("/", logoutController);
-
-export default router;
