--- backend/routes/refresh.routes.ts
@@ -1,8 +1,0 @@
-import express from "express";
-import { refreshController } from "../controllers/refreshController";
-
-const router = express.Router();
-
-router.post("/", refreshController);
-
-export default router;
