--- backend/routes/root.routes.ts
@@ -1,8 +1,0 @@
-import express, { Request, Response } from "express";
-import { rootController } from "../controllers/rootController";
-
-const router = express.Router();
-
-router.get("/", rootController);
-
-export default router;
